feat(backend): policy status codes, self-withdraw, clippy, docs
Introduce PolicyDenial (403 authz / 400 invalid state) for membership
guards. Add POST /api/applications/mine/withdraw for applicants with
shared transition path and apply-page UI. Clippy-clean app and db.
Document ban/lift semantics and archive the membership review notes.

+++ crates/app/src/pages/admin/members.rs
@@ -840,7 +840,7 @@ pub fn AdminMembers() -> Element {
                                     Some(f) => format!(
                                         "{} ({:.1} KB)",
                                         f.name(),
-                                        f.size() as f64 / 1024.0
+                                        f.size() / 1024.0
                                     ),
                                     None => "No file selected yet — click here".to_string(),
                                 };
